backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ✅ Google Gemini AI (FREE)
google-generativeai==0.7.0

# Database — works with SQLite (local) and PostgreSQL (Supabase/Railway)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9   # PostgreSQL driver for cloud
aiosqlite==0.19.0         # SQLite driver for local dev

# Market Data (FREE)
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.3
pandas-ta==0.3.14b
requests==2.31.0

# NLP (FREE)
nltk==3.8.1

# Background Tasks (no Redis needed)
apscheduler==3.10.4
cachetools==5.3.2

# Notion Integration (FREE)
notion-client==2.2.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
websockets==12.0
